perf: switch streaming fade-in from per-word to per-line
User feedback: per-word fade caused horizontal jitter inside a single
visual row because words within the same paragraph mounted at slightly
different times (deltas arrive token-by-token), so each row read as a
shimmer instead of a coherent reveal.

Switch the unit from word to line. The rehype plugin now wraps each
text node in a single <span class="sd-line"> instead of splitting on
whitespace. As text grows inside an existing span, react-markdown
reconciles in place — no remount, no re-fade. A new fade only fires
when a real boundary is crossed:
  - new paragraph (new <p> mounts → its first text-node span mounts)
  - new <br> (reasoning content with preserveLineBreaks=true runs
    remarkBreaks, which converts \n → <br>; each text node between
    breaks is a separate line span)
  - new list item, table cell, blockquote line, etc.

Renamed plugin/file/test/class for clarity:
  rehypeSplitWordsForFade → rehypeWrapLinesForFade
  .sd-word → .sd-line
  @keyframes sd-word-blur-in → sd-line-blur-in

Behavior unchanged for skip rules: <pre>/<code>/KaTeX/Mermaid subtrees
are still excluded. Blur duration (500ms) and intensity (3px) preserved.

Long single paragraphs that grow without internal newlines (the rare
case for long-form reasoning) now grow silently inside one line span —
which is a minor regression on word-by-word visibility but a clear win
on horizontal jitter, per user direction.
