--- src/browser/features/Messages/rehypeSplitWordsForFade.ts
@@ -1,116 +1,0 @@
-import type { Element, ElementContent, Root, RootContent, Text } from "hast";
-import type { Plugin } from "unified";
-
-/**
- * Rehype plugin: wrap each whitespace-bounded word in a `<span class="sd-word">`
- * inside the rendered markdown HAST. When paired with the per-word blur-fade-in
- * CSS rule in `globals.css`, this gives a soft word-by-word reveal during
- * streaming — including inside an already-mounted paragraph or thinking block,
- * which is the case the per-block fade-in alone could not address.
- *
- * Stable identity: as text grows ("Hel" → "Hello"), the same `<span>` is
- * reconciled in place by react-markdown — no remount, no re-animation. Only
- * when a fresh whitespace boundary is crossed does a new `<span>` mount, which
- * is when the CSS animation fires. That's the behavior we want.
- *
- * Skip rules:
- *  - Inside `<pre>` or `<code>` — code keeps its own monospaced/static feel.
- *  - Inside KaTeX subtrees (class contains "katex") — splitting math breaks
- *    glyph metrics. KaTeX output is generated by rehypeKatex and tagged with
- *    that class, so this plugin must run AFTER rehypeKatex in the pipeline.
- *  - Inside Mermaid `<svg>` (handled implicitly: SVG namespace nodes are
- *    structural; we only split text *children* of HTML elements).
- *  - Inside `<script>`/`<style>` — sanitize already strips these but we guard
- *    defensively.
- */
-
-const SKIP_TAGS = new Set(["pre", "code", "script", "style"]);
-const SKIP_CLASS_SUBSTRINGS = ["katex", "mermaid"];
-
-function hasSkipClass(node: Element): boolean {
-  const cls = node.properties?.className;
-  if (!Array.isArray(cls)) return false;
-  for (const entry of cls) {
-    const s = String(entry);
-    for (const needle of SKIP_CLASS_SUBSTRINGS) {
-      if (s.includes(needle)) return true;
-    }
-  }
-  return false;
-}
-
-/**
- * Split a text string into a sequence of HAST children:
- *   "Hello  world" → [<span>Hello</span>, "  ", <span>world</span>]
- *
- * Whitespace runs stay as plain Text nodes so layout (line wrapping, justify,
- * etc.) is unchanged. Empty input is a no-op.
- */
-function splitTextIntoWordSpans(value: string): ElementContent[] {
-  if (value.length === 0) return [];
-  const out: ElementContent[] = [];
-  // Match non-whitespace runs OR whitespace runs. Together this tokenizes the
-  // entire string with no gaps.
-  const re = /(\S+)|(\s+)/g;
-  let match: RegExpExecArray | null;
-  while ((match = re.exec(value)) !== null) {
-    const word = match[1];
-    const ws = match[2];
-    if (word !== undefined) {
-      const span: Element = {
-        type: "element",
-        tagName: "span",
-        properties: { className: ["sd-word"] },
-        children: [{ type: "text", value: word } satisfies Text],
-      };
-      out.push(span);
-    } else if (ws !== undefined) {
-      out.push({ type: "text", value: ws });
-    }
-  }
-  return out;
-}
-
-function isElement(node: ElementContent | RootContent): node is Element {
-  return node.type === "element";
-}
-
-function isText(node: ElementContent | RootContent): node is Text {
-  return node.type === "text";
-}
-
-function transformChildren(parent: { children: ElementContent[] | RootContent[] }): void {
-  // Re-allocate `children` so we can replace single text nodes with a sequence.
-  const next: ElementContent[] = [];
-  for (const child of parent.children as ElementContent[]) {
-    if (isText(child)) {
-      next.push(...splitTextIntoWordSpans(child.value));
-      continue;
-    }
-    if (isElement(child)) {
-      if (SKIP_TAGS.has(child.tagName) || hasSkipClass(child)) {
-        next.push(child); // do not descend
-        continue;
-      }
-      transformChildren(child);
-      next.push(child);
-      continue;
-    }
-    next.push(child);
-  }
-  parent.children = next;
-}
-
-/**
- * Standalone transform function. Exported separately from the plugin wrapper
- * so unit tests can drive it directly without spinning up a unified processor.
- */
-export function transformWordsForFade(tree: Root): void {
-  transformChildren(tree);
-}
-
-export const rehypeSplitWordsForFade: Plugin<[], Root> = () => {
-  return (tree) => {
-    transformWordsForFade(tree);
-  };
-};
