+++ src/browser/features/Messages/rehypeWrapLinesForFade.ts
@@ -1,0 +1,111 @@
+import type { Element, ElementContent, Root, RootContent, Text } from "hast";
+import type { Plugin } from "unified";
+
+/**
+ * Rehype plugin: wrap each text node in a `<span class="sd-line">`. When paired
+ * with the per-line blur-fade-in CSS rule in `globals.css`, this gives a soft
+ * line-by-line reveal during streaming.
+ *
+ * "Line" here = one HAST text node, which in practice maps to:
+ *  - One paragraph of prose (no internal `\n` because remark collapses them).
+ *  - One line between `<br>` siblings (reasoning content with
+ *    `preserveLineBreaks=true` runs `remarkBreaks`, which converts `\n` →
+ *    `<br>`; each text node between breaks is a separate line).
+ *  - One list item, table cell, blockquote line, etc.
+ *
+ * Why per-line instead of per-word: a previous iteration split each text node
+ * into N word spans and animated each. That worked, but several words within
+ * the same visual row mounted at slightly different times (because deltas
+ * arrive token-by-token), so the row read as a horizontal shimmer instead of a
+ * coherent line reveal. Per-line keeps the blur effect but eliminates the
+ * intra-row stagger: as text grows inside an existing span, react-markdown
+ * reconciles the same `<span>` in place, no remount, no re-fire. Only crossing
+ * a real boundary (new paragraph, new `<br>`, new list item) mounts a fresh
+ * span and fires the animation.
+ *
+ * Skip rules:
+ *  - Inside `<pre>` or `<code>` — code keeps its own monospaced/static feel.
+ *  - Inside KaTeX subtrees (class contains "katex") — wrapping math text
+ *    nodes corrupts glyph layout. KaTeX output is generated by rehypeKatex
+ *    and tagged with that class, so this plugin must run AFTER rehypeKatex.
+ *  - Inside Mermaid `<svg>` (handled implicitly: SVG namespace nodes are
+ *    structural; we only walk HTML element children).
+ *  - Inside `<script>`/`<style>` — sanitize already strips these, but we
+ *    guard defensively.
+ */
+
+const SKIP_TAGS = new Set(["pre", "code", "script", "style"]);
+const SKIP_CLASS_SUBSTRINGS = ["katex", "mermaid"];
+
+function hasSkipClass(node: Element): boolean {
+  const cls = node.properties?.className;
+  if (!Array.isArray(cls)) return false;
+  for (const entry of cls) {
+    const s = String(entry);
+    for (const needle of SKIP_CLASS_SUBSTRINGS) {
+      if (s.includes(needle)) return true;
+    }
+  }
+  return false;
+}
+
+/**
+ * Wrap a text node's value in a single `<span class="sd-line">`. The whole
+ * text node is one unit — no internal splitting — so growth-in-place inside
+ * the same text node does not re-mount the span.
+ */
+function wrapTextAsLineSpan(value: string): Element | null {
+  if (value.length === 0) return null;
+  return {
+    type: "element",
+    tagName: "span",
+    properties: { className: ["sd-line"] },
+    children: [{ type: "text", value } satisfies Text],
+  };
+}
+
+function isElement(node: ElementContent | RootContent): node is Element {
+  return node.type === "element";
+}
+
+function isText(node: ElementContent | RootContent): node is Text {
+  return node.type === "text";
+}
+
+function transformChildren(parent: { children: ElementContent[] | RootContent[] }): void {
+  // Re-allocate children so we can replace single text nodes with spans (or
+  // drop empty ones). Most siblings are unchanged.
+  const next: ElementContent[] = [];
+  for (const child of parent.children as ElementContent[]) {
+    if (isText(child)) {
+      const wrapped = wrapTextAsLineSpan(child.value);
+      if (wrapped) next.push(wrapped);
+      continue;
+    }
+    if (isElement(child)) {
+      if (SKIP_TAGS.has(child.tagName) || hasSkipClass(child)) {
+        next.push(child); // do not descend
+        continue;
+      }
+      transformChildren(child);
+      next.push(child);
+      continue;
+    }
+    next.push(child);
+  }
+  parent.children = next;
+}
+
+/**
+ * Standalone transform function. Exported separately from the plugin wrapper
+ * so unit tests can drive it directly without spinning up a unified processor.
+ */
+export function transformLinesForFade(tree: Root): void {
+  transformChildren(tree);
+}
+
+export const rehypeWrapLinesForFade: Plugin<[], Root> = () => {
+  return (tree) => {
+    transformLinesForFade(tree);
+  };
+};
